ldscripts: fix EH_FRAME segment
We currently place the eh_frame_hdr section in the data segment near the
eh_frame section. However, non-relinked binaries have a GNU_EH_FRAME
segment which only contains this section. Fix this by putting the
.eh_frame_hdr section in both the data (or text, as appopriate) segment
and the eh_frame_hdr segment. The eh_frame_hdr segment, unfortunately,
is not easily definable, so we have to use a constant (found by Adam) to
add it to each linker script (the constant is defined in
/usr/include/elf.h). Add some comments to indicate the lack of the
constant's definition.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>
Signed-off-by: Adam Litke <agl@us.ibm.com>