elf64ppc.B: fix eh_frame_hdr placement be with the text not data
For some reason, I made the ppc64 B-only linker script place the
eh_frame_hdr with the data, while every other arch placed it with the
text. This led to segfault of bwaves from SPECCPU in the B-only 64-bit
configuration. Fix this by placing the eh_frame_hdr in the text like all
the other scripts.

64-bit B-only bwaves completes successfully with this patch applied.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>
Acked-by: David Gibson <david@gibson.dropbear.id.au>