libhugetlbfs: Test handling of misaligned mmap() parameters
This patch adds a new testcase which checks the kernel's handling of
misaligned parameters to mmap().  That is, mmap() address, length, or
offset parameters which are not hugepage-aligned.

There have been several occasions where reworks of the kernel
get_unmapped_area() path have resulted in missing the necessary
alignment tests for hugepages.  This testcase should make such
problems easier to catch in future.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>