libhugetlbfs: Fix (sort of) the zero_filesize_segment testcase
The zero_filesize_segment testcase has been broken by changes in
binutils for quite some time.  It only ever really worked by accident,
relying on fine details of how it's custom linker script would
interact with the default linker script.

This patch improves the situation, although it's still far from
perfect:
	- it uses a new approach in the linker script which is much
simpler and more robust.  It's still not really guaranteed to work,
but the chances are much better.
	- depending on the exact details of platform and binutils
version, the old approach could result in a testcase binary which
appeared to pass, but which in fact didn't test what it was supposed
to.  We now add C code to the testcase to verify that the executable
does actually have the properties we need to test the relevant kernel
bug.
	- the testcase still won't link with some older-than-current
but still not very old ld versions.  If the testcase fails to link, we
replace it with a tiny script which just reports the failure to build
when running the testsuite.  That way systems without a sufficiently
recent binutils version can at least run the rest of the testsuite
without encountering build failures due to this hairy testcase.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>
Tested-by: Steve Fox <drfickle@linux.vnet.ibm.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>