stack_grow_into_huge: fix-up testcase for 64-bit binaries
Currently, we mmap in a huge page arbitrarily and then start growing the
stack down. For 64-bit binaries, though, the stack may be rather high up
in the address space, and there is no guarantee our huge page mapping is
anywhere near it. So, instead, try to mmap() in a hugepage very close to
the stack. If we fail, just keep moving down a hugepage at a time, until
we succeed. For x86/x86_64, this should happen relatively quickly. For
ppc64, this should put us in the segment just below the one containing
the stack.

Then, in the child process, only grow the stack down to the point where
it is below the mmap() address. If we're still running at that point,
we've clearly not been signalled.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>
Acked-by: David Gibson <david@gibson.dropbear.id.au>