elflink: Fork creation of hugetlbfs mappings.
[PPC] Prior to memory slices being added in 2.6.22, our temporary hugetlbfs
mappings are created in the segment below the stack. If a 32-bit PPC app
grows its stack larger than one segment, it will segfault because the
neighboring segment has been marked as "hugepage-only".

Providing a hint address for our temporary mappings is unreliable because,
prior to 2.6.22, if the hint address is invalid get_unmapped_area()
simplistically started at the top of the address space and searched down.
This means we could end up with the kernel picking an address that's higher
than the hint address.

To get around this limitation, David Gibson suggested forking a child to
perform the temporary hugetlbfs mappings in its own address space to avoid
tainting segments in the parent's. I have also measured a small performance
improvement, which Adam believes may be due to SLBs.

Since there is no performance hit on x86 and x86_64, nor 64-bit binaries,
perform the forking regardless of architecture or bit-ness.

Signed-off-by: Steve Fox <drfickle@linux.vnet.ibm.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>