elf64ppc.xB flexible BSS alignment
glibc makes certain assumptions about the layout of the text, data, and
bss segments of shared objects and executables.  One of those
assumptions is that all segments of an objects will be mapped
consecutively.

The current elf64ppc.xB linker script unconditionally begins the BSS at
1.5T which virtually guarantees that something will be mapped between
the data segment and bss of the executable.  This breaks the consecutive
mapping assumption and can cause application failures.

In many cases, the text, data, and bss can all fit below 4G.  In this
case, BSS alignment can be reduced to 256M and executable segments can
be mapped much closer together.  This patch contains a conditional ALIGN
statement so the case of a data segment extending beyond 4G can be
handled.

This patch resolves the dlopen problems we have been seeing lately with
some benchmark suites.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>