disable heap shrinking by default
Disable heap shrinking by default unless HUGETLB_MORECORE_SHRINK=yes is
set in the environment.

If malloc allocates memory on the heap before libhugetlbfs's
setup_morecore is called, then when malloc calls hugetlbfs_morecore it
will notice a gap between the previous top-of-heap and the memory
allocated by hugetlbfs_morecore.  Malloc treats this as a "foreign sbrk"
and marks the gap as allocated so it will avoid using this memory.
Unfortunately, it does so by calling a routine called _int_free, which
will try to shrink the heap if there is too much free memory.  Oops!
The memory allocated by hugetlbfs_morecore hasn't yet been marked as
allocated by malloc, and the heap shrinks.  Without the newly-allocated
memory, malloc returns NULL.

I've reported this as a bug to the ptmalloc2 maintainer (upstream
contributor to glibc) and provided a patch to correct the behavior in
ptmalloc2.  Unfortunately, I've not gotten a reply in two months, so it
seems prudent to disable heap shrinking in libhugetlbfs until the
behavior of glibc's malloc is changed.

Also adds test cases to trigger the unexpected malloc behavior.

Signed-off-by: Andrew Hastings <abh@cray.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>