elflink: Integrate extra-copy in prepare_segment()
Since seg->extrasz is either zero or a valid number of extra bytes to
copy, it is not necessary to test if it's greater than zero before
adding it to seg->filesz.  We can also consolidate the DEBUG message and
perform all of the copying with just one call to memcpy.

The resulting code is easier to read and integrates better with a
follow-on patch.  * Note *  I removed some confusing comments that no
longer make sense.  Follow-on patches will comment the remapping process
more concisely.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Acked-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>