libhugetlbfs: consolidate to one constructor
Use one constructor to control the constructor order for libhugetlbfs.
Currently, the constructors are run in the order their containing object
files are linked in to libhugetlbfs.so. This is fragile as new features
are added. Instead, have one constructor that calls the others (which
are now no longer actually constructors).

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>
Acked-by: Andrew Hastings <abh@cray.com>
Acked-by: David Gibson <david@gibson.dropbear.id.au>