morecore: dump /proc/pid/maps when the heap is forced to move and DEBUG
If HUGETLB_DEBUG is set in the environment, use a new helper function,
dump_proc_pid_maps() to indicate the address space layout. This is very
useful in finding out *why* the hugepage heap cannot be placed where
we'd like it. Getting this information for benchmarks, often not run
directly, helps resolve bugs.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>