elflink: filter extracopy on GLIBC symbols
I found that when using OpenMPI with libhugetlbfs there is a gigantic symbol in
the BSS. From readelf -a:

    81: 00000000120048b0 6caf2d0 OBJECT  GLOBAL DEFAULT   26	fields

This object is almost 109M in size! But because it is an OBJECT type and
GLOBAL, we copy it in the extracopy logic. However, per the comments in
elflink.c, this is not a GLIBC symbol, so we shouldn't need to. Use
strstr() to search for GLIBC in the symbol names we find, and skip those
that dont' match.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>
Acked-by: Adam Litke <agl@us.ibm.com>