Extend 'counters' test case coverage
The 'counters' test case already runs through a series of operations
that affect the hugetlb pool in measured ways.  Rather than always
starting the hugetlb pool at zero pages, if we run the tests with a base
number of pages varying from zero to the total needed, we increase
corner case test coverage.  The tests will be run with a varying number
of surplus pages used each time.

In order for this to work properly, some bugs in the kernel counters
emulation routines (_map, _unmap, and _touch) needed fixing.  _map
previously assumed there would always be a shortfall and this is no
longer true.  _unmap assumed that all unused reserved pages were also
surplus pages but some of the unused pages may now be regular pool
pages.  _touch assumed that the pool size always increases when touching
a private page, but sometimes an existing pool page can be used and the
total pool size will not change.

Signed-off-by: Adam Like <agl@us.ibm.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>