counters: reset hugepage value properly
Currently, if we do not have dynamic hugepage support in the kernel
(e.g., SLES10* or RHEL5*), we reset the static hugepage pool's size with
the value in saved_nr_hugepages. But this value is calculated *after* we
check to see if we have dynamic hugepage pool support (and we bail from
within that function, if we don't). The resetting of the pool is
performed by cleanup() which is called automatically on exit. Thus, we
write a global, unitialized value, defined to be 0, to
/proc/sys/vm/nr_hugepages. Save the size of the static pool *before*
checking the dynamic pool support to avoid shrinking the pool to 0.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>