morecore.c: eliminate use of mlock()
hugetlbfs_morecore() is currently calling mlock() presumably to
instantiate newly-mapped hugepages before returning to malloc().

However, this is ineffective for two reasons:
1.  The return value of mlock is ignored.  If there aren't enough
    hugepages available, the process may be killed later, when it tries
    to access the hugepages just added to the heap.
2.  Most distros have a default mlock limit well below the size of
    one hugepage.  (The kernel default is 8 base pages.)

This patch changes hugetlbfs_morecore to use readv() to instantiate the
newly-mapped hugepages.  If not all hugepages can be instantiated, the
new mapping is removed, and hugetlbfs_morecore returns NULL to malloc.
This allows for a graceful fallback to base pages if insufficient
hugepages are available.

This patch also adds a test case to verify that a process using
HUGETLB_MORECORE does not get killed when it exhausts the available free
hugepages, assuming sufficient free base pages are available.

Signed-off-by: Andrew Hastings <abh@cray.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>