build: Rename .S files
Due to the way our multi-target Make rules work, it is not possible for
two files to exist where the only difference in name is the file
extension.  For example: if foo.c and foo.S existed, the build system
would try and build foo.c into foo.o and then build foo.S into foo.o.
Oops.  Rather than changing the build system to remove this restriction
(which I deem is non-trivial), prefix the .S files in a sensible manner
so that .c files can be added with the former .S basenames.

Since the .S files each contain an assembly stub to make a system call,
it seems sensible to prefix these files with 'sys-'.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>