elflink: Add powerpc64-specific PLT size detection
PowerPC 64-bit binaries place their plt in the uninitialized (normally
zeroed) part of the data segment.  This placement requires extra copying
to be performed in order to preserve symbol lookups that have already
been completed by the dynamic linker.

Determining the number of additional bytes to copy is platform-specific.
Therefore, we use a new mechanism to specify an implementation for
elf64ppc.  The build system has been modified to notice $ELF{32|64}.c
files and, when found, build them into the library.

This implementation replaces the somewhat hand-wavy linker script
approach we used in the past to work around this discrepancy.

NOTE: This functionality was originally part of the script-less
relinking patch series.  Since it is a good stand-alone fix I have
broken it out and am pushing it on its own merit.  The mechanism for
supplying arch-specific functions has been changed based on a suggestion
from David Gibson.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>