morecore: offset needs to be long to handle larger page size
morecore: The offset into the mmap needs to be a long to handle larger
page sizes.  This allows the "HUGETLB_MORECORE=yes" tests to pass with
16G huge pages.

In trying to understand the code found the variables names "i" and "j"
made it difficult to understand.  Renamed "i" to "offset" to better
reflect its purpose.  Renamed "j" to "i" since it is an index.

Thanks to Adam Litke for helping me this problem.

Signed-off-by: Jon Tollefson <kniht@linux.vnet.ibm.com>
Acked-by: Andrew Hastings <abh@cray.com>
Signed-off-by: Adam Litke <agl@us.ibm.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>