tests: Fix direct test
In the direct test, we use a temporary file (/tmp/direct) along with a
hugetlbfs file to initiate direct I/O.  Since we don't want to step on another
program that may be using a file with the same name, we open /tmp/direct with
O_EXCL to ensure it doesn't already exist.  Unfortunately, we were forgetting
to unlink it again and the test would not run subsequently.  Fix it up.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>