morecore: add HUGETLB_NO_PREFAULT
Add yet another environment variable, this time to prevent the
prefaulting of hugepages via mlock() in morecore.c. Prefaulting in an
unintelligent way (interleave across all nodes, e.g.) when an
application is threaded/scheduled in an intelligent way (OpenMP) can
result in very low performance. As long as enough hugepages are
available for the system as a whole, there should be little difference
between HUGETLB_NO_PREFAULT set and unset. However, if hugepages are a
contested resource, applications may see SIGKILL if HUGETLB_NO_PREFAULT
is set.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>