Add "[hostname:pid]" to output messages
When running on multiple hosts with output coalesced, it is useful to
be able to isolate messages from individual processes.  This patch adds
the string "[hostname:pid] " to all messages emitted by libhugetlbfs.

This change uses a GNU extension of CPP that will eat a preceeding
comma when using ##__VA_ARGS__ and __VA_ARGS__ is empty.

Signed-off-by: Dean Luick <luick@cray.com>
Acked-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>