Update check_hugetlb_shm_group to properly read gid from /proc
check_hugetlb_shm_get tries to read the value from
/proc/vm/sys/hugetlb_shm_group and use that value in a comparison
against users gid.  This test will fail a majority of the time.  This
patch ensures that the correct value is used in the comparison by
converting the value read from the file from an ascii string to an
integer value.

Signed-off-by: Eric Munson <ebmunson@us.ibm.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>