elflink: Handle unaligned segment starting addresses
The current remapping algortithm assumes that each PT_LOAD elf segment
has a starting address (vaddr) that is aligned to the huge page size.
This condition was enforced by the custom linker scripts that are
packaged with libhugetlbfs.

Unaligned starting addresses can be corrected simply by aligning them
down to a huge page boundary and tracking the number of bytes by which
the address was adjusted (offset).  When copying segment data into huge
pages, the offset is used to ensure the data is copied to the proper
offset within the huge page
mapping.  An algorithmic solution to this issue eliminates a dependency on our
custom linker scripts.

Unaligned segments have increased the complexity of segment preparation
and remapping.  As a result, I felt it necessary to update the code
comments to (hopefully) more clearly represent the nature of the
calculations.

As suggested by David Gibson, I've added the ability to check that the
memory bounding a segment that needs to be aligned is empty and unused.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>