elflink: Support full segment remapping with the system linker scripts
With all of the dependencies on custom linker scripts removed, introduce
a segment remapping algorithm that is compatible with the system linker
scripts (binutils 2.18 or later).

The only thing that is strictly required for full segment remapping to
be possible is a large enough gap between the text and data segments.

Since this method does not mark program headers with the
PF_LINUX_HUGETLB flag and certain binary layouts can support multiple
remapping combinations, another method must be used to denote the
segments on which remapping should be attempted.  This patch overloads
the HUGETLB_ELFMAP environment variable in a backwards-compatible way to
enable selection of segments to be remapped.  Currently, only the string
"no" is recognized in the HUGETLB_ELFMAP variable and its effect is to
disable segment remapping.  Two new characters will now be recognized in
this variable: 'R' and 'W'.  When 'R' is present, read-only segments
(like text) will be remapped.  When 'W' is specified, writable segments
(like data/bss) will be remapped.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>