ld.hugetlbfs: Add support for relinking with system scripts
With the core of libhugetlbfs now supporting relinking without the need
for custom scripts, export the feature to users in an easy to use
manner.  For a binary to be properly linked to support this remapping
method, three things must be added to the ld command line during the
final link: -zcommon-page-size=<slice-size>,
-zmax-page-size=<slice-size>, and -lhugetlbfs.  Add a new option to
ld.hugetlbfs '--hugetlbfs-align' that will add the appropriate options
to the link command line.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>