Add hugectl source
There are currently 11 environment variables that can be used to control
the behavior of a program that is executed uses libhugetlbfs.  Rather
than expect the end user to remember all of the environment variable
names and valid values for each.  hugectl takes the idea of using a
single program as a front end for controlling a very specific target set
of applications with a large number of possible variables from numactl
and applies it to libhugetlbfs.

Submitting hugectl as a skeleton allows the community to take some time
to consider an interface that will remain mostly constant to the end
user while the underlying environment variable and library behavior
changes.

Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>