Provide a direct allocator API for huge pages
HUGETLB_MORECORE currently exists to allow glibc to back malloc() with
large pages instead of small pages. However, not all applications use
glibc malloc() nor is it always desirable to back malloc() with huge
pages. There exists a requirement that a hugepage-aware application be
able to allocate hugepages directly.

Currently, each application is expected to discover the filesystem
themselves, mmap() the file and other house-keeping tasks. libhugetlbfs
already implements much of this complex logic internally. This patch
exposes a simple API for the allocation and freeing of regions backed by
hugepages. The implementation is a little over-simplistic but can be
optimised later if and when applications perceive its performance to be
a bottleneck. The API itself should not need to change as a multi-page
aware API would be an additional rather than a replacement interface.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Acked-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>