libbugetlbfs: Test case for powerpc huge_ptep_set_wrprotect() bug
Until very recently (in fact, even now in mainline) powerpc kernels had
a bug in huge_ptep_set_wrprotect() which meant the 'huge' flag was not
passed down to pte_update() and hpte_need_flush().  This meant the hash
ptes for hugepages would not be correctly flushed on fork(), allowing
the parent to pollute the child's mapping after the fork().

This patch adds a testcase to libhugetlbfs for this behaviour, also
doing some other checking of the COW semantics over a fork().

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>
Acked-by: Adam Litke <agl@us.ibm.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>