Use MAP_NORESERVE for read-only mappings when HUGETLB_SHARE is enabled
When HUGETLB_SHARE=1, read-only text segments are copied to a file on
a hugetlbfs mount and shared between multiple processes.  However with
kernel 2.6.27 and later, the expectation is that pages will be reserved for
MAP_PRIVATE mappings as well as MAP_SHARED. This will render HUGETLB_SHARE
ineffective as huge pages will be reserved that already exist in the page
cache.  This patch disables reservations for read-only mappings when sharing
is enabled on the assumption mprotect() will never be called for remapped
read-only segments.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>