debug: cleanup and simplify specification of debugging levels
Currently we have three reporting levels ERROR, WARNING, and DEBUG.
Each duplicates the entire code to construct and emit the message even
though they only differ in prefix and level.  Clean this up by creating
a common REPORT macro which is parameterised with the level and use that
in the ERROR, WARNING, and DEBUG macros.  Additionally now that the
meaning of ERROR et al are encoded simply move these level defines to
their own header so that they can be reused by other applications with
differing output streams whilst maintaining common levels.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>
Acked-by: Adam Litke <agl@us.ibm.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>