elflink: tests: Make some library messages warnings and remove spurious '\n's
When the elflink code determines a binary is not suitably aligned for
the requested remapping mode, it emits a message to state this.  Change
the severity from ERROR to WARN since execution is able to continue, in
spite of the failure to remap segments.

Also, the ERROR and FAIL test macros append their own newline to the end
of messages.  Remove extra '\n' characters from the counters and
linkhuge_rw tests.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>