tests: Skip linkhuge tests when they are broken
Depending on the version of GNU binutils in use, the linkhuge or
linkhuge_rw tests may be known to fail and in some cases hang.  Allow
the test suite to complete and produce clean output again by skipping
tests that are known to be bad.

This patch handles the following case:

1) If an xBDT linker script is used on a system that requires the use of
the SPECIAL() linker script command for placement of the .plt and .got
sections, the resulting binary will behave erratically (function calls
will call the wrong functions) due to misplacement of the plt and got.
The solution is to not run linkhuge tests if the use of SPECIAL is
discovered in the system linker scripts.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>