tests: Handle expected differences brought on by private reservations
The private reservations feature in new kernels changes the semantics of
private huge page mappings in ways that are recognizable by the
libhugetlbfs test suite.  So far, counters and quota are effected.
Reserved private mappings now affect the pool counters in the same way
as shared mappings.  FS quota is now allocated up-front (mmap time) for
private mappings so going over quota will now cause the mmap to fail
rather than a SIGBUS at page instantiation time.

To preserve backwards compatibility, the test suite must correctly test
kernels with and without private reservation support.  To achieve this,
a function to check the running kernel's semantics is needed.  This
patch adds a function that attempts to check for private reservations.
Use of the pool counters in the manner employed is racy but since this
is a test suite and meant to be run by itself, the counters method
should work.  The function is clearly commented to warn against using it
for real programs.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>