Update large_mounts test to be compatible with multiple mount point /proc/mounts parsing
The method used for finding the proper hugetlbfs mount point has changed
with the multiple mount points patch set.  Instead of reading in 4kb at
a time, the mounts file is parsed line at a time with any extra
characters being pushed back onto the input stream.  This is
incompatible with the way that the current large mounts test overrides
read.  This patch changes the large mounts test to create a temporary
mounts file under /tmp that is padded with a fake /proc/mounts line to
be longer than 4kb.  This will be compatible with the new parsing method
because the input stream is used normally.

Signed-off-by: Eric B Munson <ebmunson@us.ibm.com>
Acked-by: Adam Litke <agl@us.ibm.com>
Acked-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>