hugectl: add support for preload
Preload is only useful when linking against an non-relinked binary
and only then when we are only attempting to back the heap.  No other
combination can usefully use preload.  Using preload with a relinked
binary is safe as long as you are only requesting backing of the heap.
Therefore unconditionally enable preload iff we are requesting only --heap.
As a failsafe allow the user to disable preload.

Note that we make use of the fact that for a standard install the
libraries will appear on the default libarary path, and that the loader
will correctly pick an appropriate library 32/64 bit based on the binary.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>
Acked-by: Mel Gorman <mel@csn.ul.ie>
Acked-by: Adam Litke <agl@us.ibm.com>
Acked-by: Eric B Munson <ebmunson@us.ibm.com>