Use MAP_NORESERVE for many read-only mappings
During preparation, a hugetlbfs file is mapped MAP_SHARED and the data
copied. At this point, the hugepages have been allocated and exist in
the page cache. The file later mapped MAP_PRIVATE which kernels after
2.6.26 will reserve hugepages in case mprotect() is called and a COW is
required.

In libhugetlbfs we know it's very highly unlikely that read-only
segments will be marked read-write. The impact is that the pages exist
in page cache but additional pages are reserved which will never be
used. This patch uses MAP_NORESERVE for all read-only mappings whose
filesz and memsz match on the assumption mprotect() and writes never
occur for these segments. This is expected to match all text segments as
it is unimaginable that a text segment would have additional pages
filled with zeros.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Acked-by: Adam Litke <agl@us.ibm.com>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>