build: Cleanup tools build
To properly support building and installing tools (hugectl, hugeedit,
hugecfg, etc), some changes to the initial method of building hugectl
are required.  This patch creates wildcard rules that can be used for
building either 32 or 64 bit versions of all tools.  Additionally, a
native wordsize is selected for each architecture so that only one
binary (of the desired default word size) is built and installed.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Acked-by: Andy Whitcroft <apw@shadowen.org>
Tested-by: Andy Whitcroft <apw@shadowen.org>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>