tools: hugeedit: Modify binaries to set default remapping behavior
The recently added relinking method makes it possible to choose from
multiple ways to remap segments at run-time via the HUGTLB_ELFMAP
environment variable.  If nothing is specified in that variable, then no
remapping will occur.  Sometimes it is desirable to set a default
remapping mode so that HUGETLB_ELFMAP does not need to be set.  This
semantic applied to the original remapping algorithm.

This patch adds a utility that can modify the PF_LINUX_HUGETLB segment
flags in the ELF haeder of a binary to control which segments will be
remapped by default.  The tool will also display this information.

[Plus fix for argc comparison by Andy W.]

Signed-off-by: Adam Litke <agl@us.ibm.com>
Tested-by: Andy Whitcroft <apw@shadowen.org>
Acked-by: Andy Whitcroft <apw@shadowen.org>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>