Fallback to smaller pages in get_huge_pages() when GHP_FALLBACK is specified
get_huge_pages() is an API for the allocation of hugepage-backed regions
of memory. In the event it returns NULL, the application is expected to
recover, possibly by using mmap() or malloc() to use base pages. This is
boiler-plate code that is an unnecessary burden on the application
developer. This patch adds GHP_FALLBACK to indicate get_huge_pages()
should use smaller pages if hugepages of the default size are not
available.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>