hugectl: add support for requesting which library to use
Add support for requesting a specific library set for preload.  This adds
the --library-path option.  If this option points to a directory containing
a libhugetlbfs library it is used, else it is assumed to be a library
prefix and both the 32 bit and 64 bit library directories are added
(where available).  By default the specific libraries installed with the
version of hugectl are used.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>
Acked-by: Eric B Munson <ebmunson@us.ibm.com>
Acked-by: Mel Gorman <mel@csn.ul.ie>