Be specific about what local symbols should not be exported
To override shmget(), it is necessary for the symbol to be
unversioned. However, all unversioned symbols are given local scope to avoid
internal functions being called accidently. This patch marks the internal-only
export functions clearly with the prefix __lh_ and then versions them to
be only of local scope. Two unused functions are simply deleted.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Acked-by: Adam Litke <agl@us.ibm.com>
Signed-off-by: Eric Munson <ebmunson@us.ibm.com>