Allow shmget() to be overridden to add the SHM_HUGETLB flag
There are applications that are not hugepage-aware but use shared memory such
as the postgres database. It is possible that the SHM_HUGETLB flag could
be added to such an application via libhugetlbfs and overriding shmget()
to add the SHM_HUGETLB to the flags. This patch enables libhugetlbfs to add
the SHM_HUGETLB when a HUGETLB_SHM environment variable is set to "yes".
It will work whether the application is linked to libhugetlbfs or loaded
via LD_PRELOAD.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Acked-by: Adam Litke <agl@us.ibm.com>
Signed-off-by: Eric Munson <ebmunson@us.ibm.com>