Fix install target for libhuge header files
When install is called with a single source file it treats the second
argument as a path ending in a file name.  So when install is called
with a single header (hugetlbfs.h) to go into /path/to/install/include,
include becomes the file name for hugetlbfs.h.  This patch changes the
call to install to force the destination to be treated as a directory and
install the listed source file(s) in that directory.

Singed-off-by: Eric B Munson <ebmunson@us.ibm.com>