tests: quota must fork and exec to change mountpoints
The multiple mounts patch introduces a slight semantic change that affects the
quota test case.  Prior to the patch, the hugetlbfs mount point was discovered
when hugetlbfs_unlinked_fd() was called for the first time.  Now the mount
points are discovered by the library constructor.  This means that the method
of setting HUGETLB_PATH and then calling hugetlbfs_unlinked_fd() will not
provide a fd from the desired mount point.

HUGETLB_PATH must be set before library initialization so change the quota test
to operate in two stages.  The first stage is to create the desired mount point
and set HUGETLB_PATH to this mount.  The second stage is entered by exec'ing
the quota test again with an argument indicating that a mount point has been
set up.  In phase two, libhugetlbfs has been allowed to initialize with the
desired mount point that was created in phase one and so the test can proceed
as it always has.

Additionally, if the libhugetlbfs default mount point has a page size that is
not the kernel default page size, the pagesize= mount option must be specified
when the quota test creates its mount point.

Signed-off-by: Adam Litke <agl@us.ibm.com>