Check HUGETLB_FEATURES for correctness
The HUGETLB_FEATURES environment variable is prone to typos and other errors
that could lead the library to behave in strange ways.  Make sure that, if
specified, the variable is checked for correctness.  The string is assumed to
be delimited by commas.  Each token is then compared against this list of known
features to ensure it matches one of them.  If no match is found, an error is
printed that explains the valid format and the whole variable is ignored.

Signed-off-by: Adam Litke <agl@us.ibm.com>